Rewrite much of the https connect to completely separate M2Crypto connect from SSL module connect. SSL with python 3 works and talks to at least a couple of SSL based servers.
This works.   Note that there are several issues at this point. 1) Have
not tested with certs. 2) There is no postconnection function in ssl
whereas there is with m2crypto so anybody who wrote something with
special functioning for cert verification will not have the verification
processed in python3 3) There is an additional post function in python 3
to check host name which we have NOT enabled at this point.