Commit a new round trip function and one extension to the mof_compiler
Extends the compiler to keep an compile-ordered list of classes
compiled.  This allows recompile to keep the original ordering.

Adds a new test that writes new mof output for the DMTF schema andd then
recompiles it.

Note that right now this does not work completely. The following known
errors exist:
1. Fails on the recompile in the very last class in the mof.
2. A number of classes do not compare because properties are different.
Many of these are because the class tomof for properties was not
outputing the default value.  See pr # 223
3. Appears that there may also be a difference in method output. Not yet
explored